log issue

+++ src/decode_cuda.cu
@@ -178,7 +178,7 @@ void decode_cuda(
     cudaMemset(gpubuf->qstring, 0, sizeof(char) * N * T);
 	checkCudaError();
 
-    const int num_state_bits = (int)log2(num_states);
+    const int num_state_bits = (int)log2((double)num_states);
     const float fixed_stay_score = options->blank_score;
     const float q_scale = options->q_scale;
     const float q_shift = options->q_shift;
